+++ server/routes/testRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const testController = require('../controllers/testController')
+
+router.get("/", testController.testServer) 
+
+module.exports = router;
