+++ server/server.js
@@ -4,25 +4,22 @@ const express = require('express');
 const connectDB = require('./db/connect');
 const cors = require('cors');
 
+// Route imports
+const testRoutes = require('./routes/testRoutes');
+const authRoutes = require('./routes/authRoutes')
+
 const app = express();
 
 // Middleware
 app.use(cors());
 app.use(express.json());
 
-// Test route
-app.get('/', (req, res) => {
-    const mongoose = require('mongoose');
-    res.json({ 
-        message: 'Expense Splitter API is running!',
-        database: mongoose.connection.readyState === 1 ? 'Connected' : 'Disconnected'
-    });
-});
 
-// Authentication routes
-app.use('/api/auth', require('./routes/auth'));
+app.use('/', testRoutes)
+app.use('/api/auth', authRoutes);
 
 const PORT = process.env.PORT || 3000;
+
 
 const start = async () => {
     try {
@@ -32,6 +29,7 @@ const start = async () => {
         app.listen(PORT, () => {
             console.log(`🚀 Server is listening at port ${PORT}...`);
         });
+
     } catch (err) {
         console.log('❌ Database connection error:', err.message);
     }
